merge/DRY scattered IntKeyLoadTests
Inclusivly merge scattered functionality for IntKeyLoadTest into single
location (picked integer_key_load_cli)
